Add SearchForManager

+++ MilmomStore.Server/Controllers/CartController.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Collections.Generic;
+using System.Drawing.Printing;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using Milmom_Service.IService;
 using Milmom_Service.Model.BaseResponse;
 using Milmom_Service.Model.Request.Cart;
@@ -25,6 +27,10 @@ namespace MilmomStore.Server.Controllers
         [HttpPost]
         public async Task<IActionResult> AddToCart([FromBody] CartRequest request)
         {
+            if(request == null)
+            {
+                return BadRequest("Please Inplement all information!");
+            }
             await _cartService.AddToCart(request.AccountId, request.ProductId);
             return Ok();
         }
@@ -32,6 +38,10 @@ namespace MilmomStore.Server.Controllers
         [HttpDelete]
         public async Task<IActionResult> RemoveFromCart([FromBody] CartRequest request)
         {
+            if (request == null)
+            {
+                return BadRequest("Please Inplement all information!");
+            }
             await _cartService.RemoveFromCart(request.AccountId, request.ProductId);
             return Ok();
         }
@@ -39,6 +49,10 @@ namespace MilmomStore.Server.Controllers
         [HttpGet("{accountId}")]
         public async Task<ActionResult<BaseResponse<CartResponse>>> GetCart(string accountId)
         {
+            if (accountId.IsNullOrEmpty())
+            {
+                return BadRequest("Please Input Id!");
+            }
             var cart = await _cartService.GetCart(accountId);
             return Ok(cart);
         }
@@ -46,6 +60,10 @@ namespace MilmomStore.Server.Controllers
         [HttpDelete("{accountId}")]
         public async Task<IActionResult> ClearCart(string accountId)
         {
+            if (accountId.IsNullOrEmpty())
+            {
+                return BadRequest("Please Input Id!");
+            }
             await _cartService.ClearCart(accountId);
             return Ok();
         }
@@ -53,6 +71,23 @@ namespace MilmomStore.Server.Controllers
         [HttpPut("{accountId}")]
         public async Task<ActionResult<BaseResponse<CartResponse>>> UpdateProductQuantityInCart(string accountId,int productId, int newQuantity)
         {
+            if (string.IsNullOrEmpty(accountId) && productId == 0)
+            {
+                // Check if pageIndex is not a valid integer
+                if (!int.TryParse(productId.ToString(), out _))
+                {
+                    return BadRequest("productId must be a valid integer.");
+                }
+
+                // Check if pageSize is not a valid integer
+                if (!int.TryParse(newQuantity.ToString(), out _))
+                {
+                    return BadRequest("newQuantity must be a valid integer.");
+                }
+
+                // Continue with your logic if all conditions are met
+                return BadRequest("Please Inplement all information!");
+            }
             var cartUpdate = await _cartService.UpdateProductQuantityInCart(accountId,productId, newQuantity);
             return Ok(cartUpdate);
         }
